SalesAPI/requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
packaging==24.1
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
whitenoise==6.7.0
